refactor!: ParticleCollection -> DATABASE (ComPWA/expertsystem#110)
Previously, the particle.DATABASE was immediately constructed from XML
through xmltodict. Now, it is passed through ParticleCollection first.
This is a step towards letting the expertsystem work with the Particle
class internally.

Note: dynamics info has to be added by the user later on because the
Particle class does not contain dynamics info!

* refactor!: remove check on NonResonant decays
* refactor!: change particle.charge to `int`
* refactor: remove Fix values from particle_list
* feat: check if spin is multiple of 0.5
* feat!: XML -> ParticleCollection -> DATABASE
* fix: remove str cast

+++ expertsystem/data.py
@@ -96,7 +96,7 @@ class Particle(NamedTuple):
 
     name: str
     pid: int
-    charge: float
+    charge: int
     spin: float
     mass: MeasuredValue
     strangeness: int = 0
